Добавлен класс дао для user и соответствующие миграции Написан тест для класса дао

+++ migrations/sql/migrations/V00.00.0002__tables.sql
@@ -4,7 +4,7 @@ CREATE TABLE tables
 (
     table_id BIGSERIAL,
     number INTEGER UNIQUE NOT NULL,
-    table_status_id INTEGER NOT NULL,
+    table_status_id BIGINT NOT NULL,
     CONSTRAINT pk_tables PRIMARY KEY (table_id),
     CONSTRAINT fk_table_statuses FOREIGN KEY (table_status_id) REFERENCES table_statuses (table_status_id)
 );
